Fix tag isolation
